// src/app/models/student.model.ts
export interface Student {
    id: number;
    name: string;
    lastName: string;
    emailAddress: string;
    documentType: string;
    documentNumber: string;
    birthDate: Date;
    sex: boolean;
    phoneNumber?: string;
    nationality: string;
    photo?: string;
    createDate: string;
  }
  